feat(ch07): LED-Roulette
My solution. It's pretty annoying that there was already a solution in
the place I'm supposed to put my own, but I didn't really give it a look
before deleting it to start over. I used the "light it all" example as
my new starting point, since I had already read that code (as it was
shown to me in the book before the challenge).

Signed-off-by: hashcatHitman <155700084+hashcatHitman@users.noreply.github.com>